dtls-server: moved helper functions to helper file, constants to constants file
