+++ sensor-server/dtls-server/dtls_server.c
@@ -12,214 +12,13 @@
 #include <signal.h>
 #include <stdbool.h>
 #include <unistd.h>
-
-
-#define SERV_PORT   20220
-#define MSGLEN      4096
-
-// STATES
-#define S_RECIEVE_REQUEST           0 // recieve sensor request
-#define S_SEND_REQUEST_TO_SERVER    1
-#define S_SEND_SENSOR_ACCEPT        2 // send sensoracc
-#define S_RECIEVE_MEASUREMENTS      3 // recieve measurements
-#define S_SEND_SENSOR_ACK           4 // send ack
-#define S_SEND_REQUEST_SUBMITED     5
-#define S_DONE                      6 
-
-// ERROR STATES
-#define ES_BAD_REQUEST_FORMAT                   -1
-#define ES_BAD_MEASUREMENT_FORMAT               -2
-#define ES_API_COMMUNICATION_BAD_REQUEST        -3 // error communicating with API
-#define ES_API_COMMUNICATION_BAD_MEASUREMENT    -4
-
-#define ES_API_CONNECTION_REQUEST               -10
-#define ES_API_CONNECTION_MEASUREMENT           -11
-
-#define ES_READ_FAILED                          -21
-#define ES_WRITE_FAILED                         -20
-
-// MESSAGES CONSTANTS
-#define SEPARATOR           '%'
-#define MIN_SREQ_LEN        29
-#define MAX_SREQ_LEN        34
-#define MAC_LEN             17
-#define MEASUREMENT_LEN     5   // measurement length - format eg T1750; - temperature 17.50
-
-// API COMMUNICATION - responses
-#define SENSOR_KNOWN                1 // sensor was already registered
-#define SENSOR_REG_SUB              2 // sensor was not registered, register request submitted
-#define BAD_REQUEST                 3 // server responed with bad request  
-
-#define MEASUREMENTS_ACCEPTED       2
-#define MEASUREMENTS_REJECTED       3
-#define ERROR_IN_API_COMMUNICATION  -1
-
-#define API_REQUEST_ENDPOINT        "http://localhost:8000/sensor/register/"
-#define API_MEASUREMENTS_ENDPOINT   "http://localhost:8000/sensor/measurements/"
+#include "constants.h"
+#include "helpers.h"
 
 static int cleanup;
-struct sockaddr_in6 servAddr;
-struct sockaddr_in6 cliaddr;
+struct sockaddr_in6 server_address;
+struct sockaddr_in6 client_address;
 
-
-bool validate_sreq(const char sreq[], int len){
-    const char *prefix = "SENSORREQ";
-    int prefix_len = strlen(prefix);
-    int mac_start = prefix_len + 1;
-    int suffix_start = mac_start + MAC_LEN + 1;
-
-    if (memcmp(sreq, prefix, prefix_len) != 0) {
-        return false;
-    }
-
-    if (sreq[prefix_len] != SEPARATOR || sreq[mac_start + MAC_LEN] != SEPARATOR) {
-        return false;
-    }
-
-    for (int i = 0; i < 17; i++) {
-        char c = sreq[mac_start + i];
-        if (i % 3 == 2) {
-            if (c != ':') 
-                return false;
-        } else  if (!isxdigit(c)){
-            return false;
-        }
-    }
-
-    for (int i = suffix_start; i < len; i++) {
-        if (!isupper(sreq[i]))
-        // if (!isupper(sreq[i]) && sreq[i] != '\n') // for now - to test working with openssl client
-            return false;
-    }
-
-    return true;
-}
-
-bool validate_measurements(const char ms[], int ms_len){
-    char measurement[MEASUREMENT_LEN + 1];
-    int i = 0;
-
-    if (ms_len < MEASUREMENT_LEN)
-        return false;
-
-    while (i < ms_len) {
-        int j = 0;
-
-        while (i < ms_len && ms[i] != '%' && j < MEASUREMENT_LEN) {
-            measurement[j++] = ms[i++];
-        }
-        measurement[j] = 0;
-
-        if (j != 5 || !isupper(measurement[0]))
-            return false;
-        for (int k = 1; k < MEASUREMENT_LEN; k++) {
-            if (!isdigit(measurement[k]))
-            // if (!isdigit(measurement[k]) && measurement[k] != '\n') // for development - communication with openssl client
-                return false;
-        }
-
-        if (i < ms_len && ms[i] == '%')
-            i++;
-    }
-
-    return true;
-}
-
-bool check_sensor_err(WOLFSSL* ssl) {
-    int readErr = wolfSSL_get_error(ssl, 0);
-    return readErr != SSL_ERROR_WANT_READ;
-}
-
-struct MemoryStruct {
-    char *memory;
-    size_t size;
-};
-
-static size_t WriteMemoryCallback(void *contents, size_t size, size_t nmemb, void *userp) {
-    size_t realsize = size * nmemb;
-    struct MemoryStruct *mem = (struct MemoryStruct *)userp;
-
-    char *ptr = realloc(mem->memory, mem->size + realsize + 1);
-    if(ptr == NULL) {
-        // out of memory!
-        printf("not enough memory (realloc returned NULL)\n");
-        return 0;
-    }
-
-    mem->memory = ptr;
-    memcpy(&(mem->memory[mem->size]), contents, realsize);
-    mem->size += realsize;
-    mem->memory[mem->size] = 0;
-
-    return realsize;
-}
-
-int send_request_to_api(CURL* curl, char sreq[], int len) {
-    CURLcode res;
-    int result;
-
-    struct MemoryStruct chunk;
-    chunk.memory = malloc(1);
-    chunk.size = 0; 
-    
-    printf("sending sesnor request to api: %s\n", sreq);
-
-    curl_easy_setopt(curl, CURLOPT_URL, API_REQUEST_ENDPOINT);
-    curl_easy_setopt(curl, CURLOPT_POST, 1L);
-    curl_easy_setopt(curl, CURLOPT_POSTFIELDS, sreq);
-    curl_easy_setopt(curl, CURLOPT_WRITEFUNCTION, WriteMemoryCallback);
-    curl_easy_setopt(curl, CURLOPT_WRITEDATA, (void *)&chunk);
-
-    res = curl_easy_perform(curl);
-
-    if(res != CURLE_OK) {
-        fprintf(stderr, "curl_easy_perform() failed: %s\n", curl_easy_strerror(res));
-        result = ERROR_IN_API_COMMUNICATION;
-    } else {
-        printf("sent sensor request to api: %s\n", sreq);
-        result = atoi(chunk.memory);
-    }
-
-    free(chunk.memory);
-    return result;
-}
-
-int send_ms_to_api(CURL* curl, char mac[], char ms[], int ms_len) {
-    CURLcode res;
-    int result;
-
-    struct MemoryStruct chunk;
-    chunk.memory = malloc(1);
-    chunk.size = 0; 
-
-    char msg_to_api[MAC_LEN + ms_len + 2];
-    // build message to API
-    strncpy(msg_to_api, mac, MAC_LEN);
-    msg_to_api[MAC_LEN] = SEPARATOR;
-    msg_to_api[MAC_LEN+1] = '\0';
-    strncat(msg_to_api, ms, ms_len);
-
-    msg_to_api[MAC_LEN + ms_len + 2] = '\0';
-    printf("sending measurements to API server: %s\n", msg_to_api);
-
-    curl_easy_setopt(curl, CURLOPT_URL, API_MEASUREMENTS_ENDPOINT);
-    curl_easy_setopt(curl, CURLOPT_POST, 1L);
-    printf("msg to api: %s\n", msg_to_api);
-    curl_easy_setopt(curl, CURLOPT_POSTFIELDS, msg_to_api);
-    curl_easy_setopt(curl, CURLOPT_WRITEFUNCTION, WriteMemoryCallback);
-    curl_easy_setopt(curl, CURLOPT_WRITEDATA, (void *)&chunk);
-
-    res = curl_easy_perform(curl);
-
-    if(res != CURLE_OK) {
-        fprintf(stderr, "curl_easy_perform() failed: %s\n", curl_easy_strerror(res));
-    } else {
-        result = atoi(chunk.memory);
-    }
-
-    free(chunk.memory);
-    return result;
-}
 
 int handle_client(WOLFSSL* ssl, CURL* curl) {
     int recv_len    = 0;
@@ -465,11 +264,11 @@ int main(int argc, char** argv) {
         }
         printf("info: socket allocated\n");
 
-        /* clear servAddr each loop */
-        memset((char *)&servAddr, 0, sizeof(servAddr));
+        /* clear server_address each loop */
+        memset((char *)&server_address, 0, sizeof(server_address));
 
-        servAddr.sin6_family      = AF_INET6;
-        servAddr.sin6_port        = htons(SERV_PORT);
+        server_address.sin6_family      = AF_INET6;
+        server_address.sin6_port        = htons(SERV_PORT);
 
         /* Eliminate socket already in use error */
         if (setsockopt(listenfd, SOL_SOCKET, SO_REUSEADDR, &on, len) < 0) {
@@ -479,7 +278,7 @@ int main(int argc, char** argv) {
         }
 
         /*Bind Socket*/
-        if (bind(listenfd, (struct sockaddr*)&servAddr, sizeof(servAddr)) < 0) {
+        if (bind(listenfd, (struct sockaddr*)&server_address, sizeof(server_address)) < 0) {
             printf("Bind failed.\n");
             cleanup = 1;
             cont = 1;
@@ -487,17 +286,17 @@ int main(int argc, char** argv) {
 
         printf("Awaiting client connection on port %d\n", SERV_PORT);
 
-        cliLen = sizeof(cliaddr);
+        cliLen = sizeof(client_address);
         connfd = (int)recvfrom(listenfd, (char *)&b, sizeof(b), MSG_PEEK,
-                (struct sockaddr*)&cliaddr, &cliLen);
+                (struct sockaddr*)&client_address, &cliLen);
 
         if (connfd < 0) {
             printf("No clients in que, enter idle state\n");
             continue;
         }
         else if (connfd > 0) {
-            if (connect(listenfd, (const struct sockaddr *)&cliaddr,
-                        sizeof(cliaddr)) != 0) {
+            if (connect(listenfd, (const struct sockaddr *)&client_address,
+                        sizeof(client_address)) != 0) {
                 printf("Udp connect failed.\n");
                 cleanup = 1;
                 cont = 1;
